refactor(summary): split reader artifact queries
